Add plugin crash example
